Update knn logic and execution
Add CMakeLists file, update knn main function, add openmp parallelism, decouple knn logic into different functions, update class DataPoint predicted classes attribute type, update gitignore to not include build directory

+++ knn.hpp
@@ -9,7 +9,7 @@ class DataPoint
     vector<float> fields;
     int size;
     string class_name;
-    pair<vector<int>, vector<double>> predicted_classes;
+    pair<vector<int>, vector<float>> predicted_classes;
 
     DataPoint(void) {
         size = 4;
@@ -19,9 +19,9 @@ class DataPoint
     void print(int main_idx, int field_length) {
         printf("\nDataPoint [%d] | ", main_idx);
         printf("Size: %d | ", size);
-        printf("Class: %d | ", fields[4]);
+        printf("Class: %f | ", fields[4]);
         for (int i = 0; i < field_length; i++) {
-            printf("%d ", fields[i]);
+            printf("%f ", fields[i]);
         }        
     }
